+++ internal/database/migration/cliutil/validate.go
@@ -11,9 +11,10 @@ import (
 
 func Validate(commandName string, factory RunnerFactory, outFactory OutputFactory) *cli.Command {
 	schemaNamesFlag := &cli.StringSliceFlag{
-		Name:  "db",
-		Usage: "The target `schema(s)` to validate. Comma-separated values are accepted. Supply \"all\" to validate all schemas.",
-		Value: cli.NewStringSlice("all"),
+		Name:    "schema",
+		Usage:   "The target `schema(s)` to validate. Comma-separated values are accepted. Possible values are 'frontend', 'codeintel', 'codeinsights' and 'all'.",
+		Value:   cli.NewStringSlice("all"),
+		Aliases: []string{"db"},
 	}
 	skipOutOfBandMigrationsFlag := &cli.BoolFlag{
 		Name:  "skip-out-of-band-migrations",
@@ -22,7 +23,7 @@ func Validate(commandName string, factory RunnerFactory, outFactory OutputFactor
 	}
 
 	action := makeAction(outFactory, func(ctx context.Context, cmd *cli.Context, out *output.Output) error {
-		schemaNames := sanitizeSchemaNames(schemaNamesFlag.Get(cmd))
+		schemaNames := sanitizeSchemaNames(schemaNamesFlag.Get(cmd), out)
 		if len(schemaNames) == 0 {
 			return flagHelp(out, "supply a schema via -db")
 		}